refactor(channel-search): extract private helpers into internal.ts
Move type mappers and normalizers (cleanFilters / secondsToDateOnly /
normalizeItems / senderFromHit / channelFromHit / mapMessageHit /
mapFileHit / mapMediaHit / mapCombinedHit / mapMessageMediaHit /
normalizeRichText / mapForwardInnerMessage / parentGroupChannel /
hasEffectiveFilters / normalizeImageUrl / normalizeFileUrl / ...) from
apiAdapter.ts into a new internal.ts as named exports so the upcoming
GlobalSearch data source can reuse them. Behavior unchanged; existing
ChannelSearch tests continue to pass (52/52).

Prep for feat(global-search) v10 P1.
